oltp_generator/requirements.txt
pymssql==2.2.5
Faker==18.9.0
python-dotenv==1.1.0

